requirements.txt
chainchomplib~=0.0.10
flask
PyYAML
Jinja2~=2.11.2
asyncio
setuptools~=51.1.1
aiohttp~=3.7.3
python-dispatch
python-socketio~=5.0.2